:sparkles: implement postprocessing
