+++ tracex/extraction/logic/modules/module_metrics_analyzer.py
@@ -87,10 +87,5 @@ class MetricsAnalyzer(Module):
         timestamp_correctness, top_logprops = u.query_gpt(
             messages, logprobs=True, top_logprobs=1
         )
-        linear_prop = self.__calculate_linear_probability(top_logprops[0].logprob)
+        linear_prop = u.calculate_linear_probability(top_logprops[0].logprob)
         return (timestamp_correctness, linear_prop)
-
-    @staticmethod
-    def __calculate_linear_probability(logprob):
-        linear_prob = np.round(np.exp(logprob), 2)
-        return linear_prob
