+++ tracex/extraction/logic/prompts.py
@@ -329,100 +329,6 @@ COMPARE_MESSAGES = [
         "content": "First: experiencing first covid 19 symptoms\nSecond: experiencing worse symptoms",
     },
     {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: putting loved ones over financial worries\nSecond: consulting a doctor",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: isolating myself\nSecond: beginning self isolation",
-    },
-    {"role": "assistant", "content": "True"},
-    {
-        "role": "user",
-        "content": "First: informing family\nSecond: informing family about my condition",
-    },
-    {"role": "assistant", "content": "True"},
-    {
-        "role": "user",
-        "content": "First: getting hospitalized\nSecond: consulting family physician",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: taking immediate action\nSecond: isolating myself",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: isolating at home\nSecond: experiencing first covid 19 symptoms",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: advised home quarantine and treatment\nSecond: consulting family physician",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: informing family about situation\nSecond: isolating myself",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: consulting family physician\nSecond: consulting doctor",
-    },
-    {"role": "assistant", "content": "True"},
-    {"role": "user", "content": "First: isolating myself\nSecond: isolating at home"},
-    {"role": "assistant", "content": "True"},
-    {"role": "user", "content": "First: informing family\nSecond: isolating myself"},
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: returning to work as teacher\nSecond: returning to work with precautions",
-    },
-    {"role": "assistant", "content": "True"},
-    {
-        "role": "user",
-        "content": "First: grateful for family and healthcare support\nSecond: feeling thankful for healthcare personal",
-    },
-    {"role": "assistant", "content": "True"},
-    {
-        "role": "user",
-        "content": "First: stabilizing condition without hospitalization\nSecond: monitoring oxygen levels",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: joining virtual support groups\nSecond: focusing on self-care",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: getting the first dose of vaccination\nSecond: focusing on self-care",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: self-isolating to protect patients\nSecond: starting to experience symptoms",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: consulting doctor virtually\nSecond: worsening symptoms",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: taking prescribed medication\nSecond: stabilizing condition without hospitalization",
-    },
-    {"role": "assistant", "content": "False"},
-    {
-        "role": "user",
-        "content": "First: feeling reliefed and protected\nSecond: receiving Covid-19 vaccines",
-    },
-    {"role": "assistant", "content": "False"},
 ]
 
 COHORT_TAG_MESSAGES = [
