+++ tracex/tracex/logic/utils.py
@@ -10,6 +10,7 @@ import functools
 import warnings
 import pandas as pd
 import pm4py
+import numpy as np
 
 from pandas.api.types import is_datetime64_any_dtype as is_datetime
 from django.conf import settings
@@ -134,6 +135,11 @@ def append_csv():
         for row in content:
             row = row.replace(row[0], str(int(row[0]) + trace_count), 1)
             f.writelines(row)
+
+
+def calculate_linear_probability(logprob):
+    linear_prob = np.round(np.exp(logprob), 2)
+    return linear_prob
 
 
 class Conversion:
